requirements.txt
langchain>=0.0.335
langchain-core>=0.1.7
langchain-openai>=0.0.1
langgraph>=0.0.20
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
ipython>=8.0.0
python-dotenv>=1.0.0 